Refactored the filter functions:
Instead of reading the file within the function, i've refactored the filter functions so that they take in a list as a parameter. The idea being the the list being provided comes from the transactions function. So now only one function opens and reads a file in order to prevent the same lines of code to be repeated.

To do:
add doc strings